inicio box e correçoes

+++ Phase1/Generator/main.cpp
@@ -2,6 +2,7 @@
 #include <stdlib.h>
 
 #include "figures/plane.h"
+#include "figures/box.h"
 
 
 int main(int argc, char** argv) {
@@ -11,9 +12,9 @@ int main(int argc, char** argv) {
         }
 
         if(strcmp(argv[1],"box")==0){
-            box(argv[4],atof(argv[2]), atof(argv[3]));
+            box(argv[4], atof(argv[2]), atoi(argv[3]));
         }
-
+/*
         if(strcmp(argv[1],"sphere")==0){
             sphere(argv[5],atof(argv[2]),atoi(argv[3]),atoi(argv[4]));
         }
@@ -22,7 +23,7 @@ int main(int argc, char** argv) {
             cone(argv[6],atof(argv[2]),atof(argv[3]),atoi(argv[4]),atoi(argv[5]));
 
         }
-
+*/
     return 0;
 }
